Rename bangnotation to reset
