+++ domains-Task/.jvm/src/test/scala/com/thoughtworks/dsl/domains/UsingSpec.scala
@@ -2,7 +2,7 @@ package com.thoughtworks.dsl
 package domains
 
 import Continuation.!!
-import com.thoughtworks.dsl.bangnotation._
+import com.thoughtworks.dsl.reset, reset._
 import com.thoughtworks.dsl.keywords.{Using, Yield}
 import org.scalatest.Assertion
 import org.scalatest.freespec.AnyFreeSpec
