+++ domains-scalaz/src/main/scala/com/thoughtworks/dsl/domains/scalaz.scala
@@ -26,7 +26,7 @@ import scala.util.control.NonFatal
   *          import _root_.scalaz.Free.Trampoline
   *          import com.thoughtworks.dsl.keywords.Monadic
   *          import com.thoughtworks.dsl.domains.scalaz.given
-  *          import com.thoughtworks.dsl.bangnotation.reset
+  *          import com.thoughtworks.dsl.reset
   *          import com.thoughtworks.dsl.keywords.Monadic.unary_!
   *
   *          val trampoline3 = Trampoline.done(3)
