+++ keywords-Using/src/main/scala/com/thoughtworks/dsl/keywords/Using.scala
@@ -1,7 +1,7 @@
 package com.thoughtworks.dsl
 package keywords
 
-import com.thoughtworks.dsl.bangnotation._
+import com.thoughtworks.dsl.reset, reset._
 import com.thoughtworks.dsl.Dsl
 import com.thoughtworks.dsl.Dsl.!!
 import com.thoughtworks.dsl.Dsl.AsKeyword
@@ -54,7 +54,7 @@ object Using {
     *           import scala.concurrent.Future
     *           import com.thoughtworks.dsl.keywords.Using.scopeExit
     *           import com.thoughtworks.dsl.keywords.Using.unary_!
-    *           import com.thoughtworks.dsl.bangnotation._
+    *           import com.thoughtworks.dsl.reset, reset._
     *           var n = 1
     *           def multiplicationAfterAddition = *[Future] {
     *             !scopeExit { () =>
