+++ keywords-Yield/src/main/scala/com/thoughtworks/dsl/keywords/Yield.scala
@@ -18,13 +18,13 @@ import scala.language.higherKinds
   *          or it will not compile.
   *
   *          {{{
-  *          import com.thoughtworks.dsl.bangnotation._
+  *          import com.thoughtworks.dsl.reset, reset._
   *          "def f(): Int = !Yield(1)" shouldNot compile
   *          }}}
   *
   * @example [[Yield]] keywords can be used together with other keywords.
   *          {{{
-  *          import com.thoughtworks.dsl.bangnotation._
+  *          import com.thoughtworks.dsl.reset, reset._
   *          def gccFlagBuilder(sourceFile: String, includes: String*) = reset[Stream[String]] {
   *            !Yield("gcc")
   *            !Yield("-c")
