+++ keywords-Yield/src/test/scala/com/thoughtworks/dsl/keywords/YieldSpec.scala
@@ -1,6 +1,6 @@
 package com.thoughtworks.dsl
 package keywords
-import bangnotation._
+import reset._
 import com.thoughtworks.dsl.Dsl.!!
 import scala.annotation.tailrec
 import scala.collection.{LinearSeq, SeqView}
