+++ reset/src/test/scala/com/thoughtworks/dsl/keywords/ReturnSpec.scala
@@ -1,7 +1,7 @@
 package com.thoughtworks.dsl
 package keywords
 import com.thoughtworks.dsl.Dsl.!!
-import bangnotation.{_, given}
+import reset.{_, given}
 import utest.{TestSuite, Tests, given}
 import Dsl.Run
 import scala.language.implicitConversions
