+++ scala-async/src/main/scala/scala/async/Async.scala
@@ -1,6 +1,6 @@
 package scala.async
 import scala.concurrent.Future
-import com.thoughtworks.dsl.bangnotation.`*`
+import com.thoughtworks.dsl.reset.`*`
 import com.thoughtworks.dsl.keywords.Await
 
 /** Async blocks provide a direct means to work with
